# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Produtos",
        "description": "API para realizar CRUD de Produtos.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:3000/"
        }
    ],
    "components": {
        "schemas": {
            "Produto": {
                "type": "object",
                "properties": {
                    "nome": {
                        "type": "string",
                        "example": "Mouse Optico Vertical Ergonomico Sem Fio"
                    },
                    "descricao": {
                        "type": "string",
                        "example": "Mouse Optico Vertical Ergonomico Sem Fio - 5 Teclas destro"
                    },
                    "quantidade": {
                        "type": "number",
                        "example": 10
                    },
                    "preco": {
                        "type": "number",
                        "example": 63.93
                    },
                    "desconto": {
                        "type": "number",
                        "example": 10
                    },
                    "dataDesconto": {
                        "type": "string",
                        "example": "2023-02-01",
                        "format": "date"
                    },
                    "categoria": {
                        "type": "string",
                        "example": "Informática"
                    },
                    "imagem": {
                        "type": "string",
                        "format": "binary",
                        "example": "Upload de imagem"
                    }
                },
                "required": [
                    "nome"
                ]
            }
        }
    },
    "paths": {
        "/produtos": {
            "post": {
                "summary": "Cadastro de produto",
                "description": "Essa rota será responsável por cadastrar um novo produto.",
                "tags": ["Produtos"],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Produto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produto adicionado."
                    },
                    "404": {
                        "description": "Produto não encontrado."
                    },
                    "500": {
                        "description": "Um erro aconteceu"
                    }
                }
            },
            "get": {
                "description": "Listar preços.",
                "summary": "Listagem de produtos",
                "tags": [
                    "Produtos"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "ID do produto a ser procurado.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produto encontrado."
                    },
                    "404": {
                        "description": "Produto não encontrado."
                    },
                    "500": {
                        "description": "Um erro aconteceu"
                    }
                }
            }
        },
        "/produtos/{id}": {
            "put": {
                "summary": "Atualização de produto",
                "description": "Atualizar um produto existente.",
                "tags": [
                    "Produtos"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id do produto.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Produto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produto atualizado."
                    },
                    "404": {
                        "description": "Produto não encontrado."
                    },
                    "500": {
                        "description": "Um erro aconteceu."
                    }
                }
            },
            "delete": {
                "summary": "Exclusão de produto",
                "description": "Excluir um produto existente.",
                "tags": [
                    "Produtos"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id do produto a ser excluído.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produto excluído."
                    },
                    "404": {
                        "description": "Produto não encontrado."
                    },
                    "500": {
                        "description": "Um erro aconteceu."
                    }
                }
            }
        }
    }
}